0.6.0 & 0.6.1: Sistema de arredondamento++

+++ src/br/com/felipedevbino/dadosgerais/ModeloParaDados.java
@@ -1,12 +1,13 @@
 package br.com.felipedevbino.dadosgerais;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public class ModeloParaDados {
-
+	
 	/**
-	 * Método para deletar itens de ArrayLists de Strings.
+	 * Método para deletar itens de uma List de Strings.
 	 * 
 	 * @param dadoParaRemocao
 	 * @param dados
@@ -22,15 +23,15 @@ public class ModeloParaDados {
 	}
 
 	/**
-	 * Método para deletar itens de HashMaps (String, List de Strings) & (String,
-	 * Double).
+	 * Método para deletar itens de Maps (String, List de Strings) & (String,
+	 * BigDecimal).
 	 * 
 	 * @param dadoParaRemocao
 	 * @param dados
 	 * @return
 	 */
 	public void remover(String dadoParaRemocao, Map<String, List<String>> dadosStr,
-			Map<String, Double> dadosDouble) {
+			Map<String, BigDecimal> dadosDouble) {
 		if (dadosStr.containsKey(dadoParaRemocao)) {
 			dadosStr.remove(dadoParaRemocao);
 
@@ -40,20 +41,20 @@ public class ModeloParaDados {
 	}
 
 	/**
-	 * Método para deletar itens de HashMaps (String, Map<String, Double>).
+	 * Método para deletar itens de Maps (String, Map<String, BigDecimal>).
 	 * 
 	 * @param dadoParaRemocao
 	 * @param dadosDoubleMap
 	 * @return
 	 */
-	public void remover(String dadoParaRemocao, Map<String, Map<String, Double>> dadosDoubleMap) {
+	public void remover(String dadoParaRemocao, Map<String, Map<String, BigDecimal>> dadosDoubleMap) {
 		if (dadosDoubleMap.containsKey(dadoParaRemocao)) {
 			dadosDoubleMap.remove(dadoParaRemocao);
 		}
 	}
 
 	/**
-	 * Método para renomear itens de ArrayLists de Strings.
+	 * Método para renomear itens de uma List de Strings.
 	 * 
 	 * @param novoNome
 	 * @param dadoParaRenomear
@@ -71,15 +72,15 @@ public class ModeloParaDados {
 	}
 
 	/**
-	 * Método para renomear itens de HashMaps (String, List de Strings) & (String,
-	 * Double).
+	 * Método para renomear itens de Maps (String, List de Strings) & (String,
+	 * BigDecimal).
 	 * 
 	 * @param novoItem
 	 * @param dadoParaRenomear
 	 * @param dados
 	 */
 	public void renomear(String novoNome, String dadoParaRenomear, Map<String, List<String>> dadosStr,
-			Map<String, Double> dadosDouble) {
+			Map<String, BigDecimal> dadosDouble) {
 
 		if (dadosStr.containsKey(dadoParaRenomear)) {
 			List<String> copiaValor = dadosStr.get(dadoParaRenomear);
@@ -87,7 +88,7 @@ public class ModeloParaDados {
 			dadosStr.put(novoNome, copiaValor);
 
 		} else if (dadosDouble.containsKey(dadoParaRenomear)) {
-			double copiaValor = dadosDouble.get(dadoParaRenomear);
+			BigDecimal copiaValor = dadosDouble.get(dadoParaRenomear);
 			dadosDouble.remove(dadoParaRenomear);
 			dadosDouble.put(novoNome, copiaValor);
 
@@ -95,22 +96,22 @@ public class ModeloParaDados {
 	}
 	
 	/**
-	 * Método para renomear itens de HashMaps (String, Map<String, Double>).
+	 * Método para renomear itens de Maps (String, Map<String, BigDecimal>).
 	 * 
 	 * @param novoNome
 	 * @param dadoParaRenomear
 	 * @param dadosDoubleMap
 	 */
-	public void renomear(String novoNome, String dadoParaRenomear, Map<String, Map<String, Double>> dadosDoubleMap) {
+	public void renomear(String novoNome, String dadoParaRenomear, Map<String, Map<String, BigDecimal>> dadosDoubleMap) {
 		if(dadosDoubleMap.containsKey(dadoParaRenomear)) { 
-			Map<String, Double> dadosDoubleMapCopia = dadosDoubleMap.get(dadoParaRenomear);
+			Map<String, BigDecimal> dadosDoubleMapCopia = dadosDoubleMap.get(dadoParaRenomear);
 			dadosDoubleMap.remove(dadoParaRenomear);
 			dadosDoubleMap.put(novoNome, dadosDoubleMapCopia);
 		}
 	}
 
 	/**
-	 * Método para modificar itens de um HashMap de (Object, String).
+	 * Método para modificar itens de um Map de (Object, String).
 	 * 
 	 * @param instancia
 	 * @param novaDificuldade
